# api-docs.json
{"openapi":"3.0.1","info":{"title":"IO Client API","description":"This is an API client that can be used to send a “message” to a citizen using the “IO” REST API.","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Message","description":"the Message API"}],"paths":{"/v1/submitMessageForUser":{"post":{"tags":["Message"],"summary":"Send a message to a user","description":"Send a custom message to a specific user identified by fiscal code","operationId":"submitMessageForUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitMessageRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"Message":{"required":["markdown","subject"],"type":"object","properties":{"subject":{"maxLength":120,"minLength":10,"type":"string","description":"Subject of the message"},"markdown":{"maxLength":10000,"minLength":80,"type":"string","description":"Text of the message in markdown"}},"description":"Message to send"},"SubmitMessageRequest":{"required":["fiscalCode","message"],"type":"object","properties":{"message":{"$ref":"#/components/schemas/Message"},"fiscalCode":{"pattern":"[A-Z]{6}[0-9LMNPQRSTUV]{2}[ABCDEHLMPRST][0-9LMNPQRSTUV]{2}[A-Z][0-9LMNPQRSTUV]{3}[A-Z]","type":"string","description":"User's fiscal code"}},"description":"Message and user"},"ErrorResponse":{"type":"object","properties":{"cause":{"type":"string","description":"Cause of the error"},"details":{"type":"string","description":"Details of the error"}}},"MessageResponse":{"required":["id"],"type":"object","properties":{"id":{"type":"string","description":"Message id"}}}}}}